On branch: master
 Changes to be committed:
	modified:   MP3Edit/BuildInfo.h
	modified:   MP3Edit/BuildNumber.h
	modified:   MP3Edit/ControlVars.h
	modified:   MP3Edit/FileDlgFactory.cpp
	modified:   MP3Edit/FileDlgFactory.h
	modified:   MP3Edit/GuiClasses/GetControlsVector.cpp
	modified:   MP3Edit/GuiClasses/SControl.cpp
	modified:   MP3Edit/GuiClasses/SDimensions.h
	modified:   MP3Edit/MP3Edit.cpp
	modified:   MP3Edit/MP3Edit.h
	modified:   MP3Edit/MP3Edit.vcxproj
	modified:   MP3Edit/MP3Edit.vcxproj.filters
	modified:   MP3Edit/StdAfx.hpp
	new file:   MP3Edit/TagsIO.cpp
	new file:   MP3Edit/TagsIO.hpp
	modified:   MP3Edit/data/Strings.hpp
	modified:   MP3Edit/genres/GenreList.cpp
	modified:   MP3Edit/genres/GenreList.hpp
	modified:   MP3Edit/libs/id3v2lib/src/utils.c
	modified:   MP3Edit/res/MP3Edit.rc
	modified:   MP3Edit/res/Resources.h
	deleted:    MP3Edit/res/icons/MP3Edit.ico
	new file:   MP3Edit/res/icons/app_icon.ico
	modified:   MP3Edit/version.txt
	modified:   README.md

+++ MP3Edit/libs/id3v2lib/src/utils.c
@@ -169,12 +169,10 @@ void println_utf16(uint16_t* string, int32_t size) {
 
 
 char* get_path_to_file(const char* file) {
-	char*    file_name = strrchr(file, '/');
-	
-	// Add 1 to account for the trailing '/':
-	uint32_t size      = strlen(file) - strlen(file_name) + 1; 
-	char*    file_path = (char*) malloc(size * sizeof(char));
+	char*  file_name = strrchr(file, '/');
+	// +1 to account for the trailing '/':
+	size_t size      = (strlen(file) - strlen(file_name) + 1);
+	char*  file_path = (char*) malloc(size * sizeof(char));
 	strncpy(file_path, file, size);
-	
 	return file_path;
 }
